fix compiler warnings on 32-bit platforms
git-svn-id: https://qthreads.googlecode.com/svn/trunk@1586 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ test/time_producerconsumer.c
@@ -169,7 +169,7 @@ int main(int argc, char *argv[])
 	qthread_empty(NULL, &(FEBtable[i][1]));
     }
     printf("Testing producer/consumer (with %i sheps):\n",
-	   qthread_num_shepherds());
+	   (int)qthread_num_shepherds());
 
     /* SINGLE FEB SEND/RECEIVE TEST */
     printf("\tSingle FEB send/receive: ");
